+++ src/blcmm/gui/panels/ObjectExplorerPanel.java
@@ -42,6 +42,7 @@ import blcmm.gui.components.FontInfoJLabel;
 import blcmm.gui.components.FontInfoJToggleButton;
 import blcmm.gui.components.InfoLabel;
 import blcmm.gui.text.AutoCompleteAttacher;
+import blcmm.gui.text.CustomSelectionMouseAdapter;
 import blcmm.gui.text.HighlightedTextArea;
 import blcmm.gui.theme.ThemeManager;
 import blcmm.utilities.CodeFormatter;
@@ -178,6 +179,9 @@ public class ObjectExplorerPanel extends javax.swing.JPanel {
 
         // Update our UI based on the current game selection
         this.updateGame(true);
+
+        // Double-click on our query textbox should select the entire word
+        queryTextField.addMouseListener(new CustomSelectionMouseAdapter(queryTextField));
 
         // Attach some handlers to various UI elements
         queryTextField.getActionMap().put("Search", new AbstractAction("Search") {